Cleaned the code;
