+++ embodichain/gen_sim/prompt2scene/agent_tools/managers/simready_manager/schemas.py
@@ -18,12 +18,22 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from pathlib import Path
+from typing import Any
+
+__all__ = [
+    "EstimateMetricScalesRequest",
+    "EstimateMetricScalesResult",
+    "GlobalMetricScaleRequest",
+    "MakeAssetSimreadyRequest",
+    "MakeAssetSimreadyResult",
+    "MakeTableSimreadyRequest",
+    "MakeTableSimreadyResult",
+    "MetricScaleObjectInput",
+]
 
 
 @dataclass(frozen=True)
 class MakeAssetSimreadyRequest:
-    """Request to prepare a general asset GLB for simulation placement."""
-
     input_path: Path
     output_path: Path
     input_up_axis: list[float] | None = None
@@ -33,16 +43,12 @@ class MakeAssetSimreadyRequest:
 
 @dataclass(frozen=True)
 class MakeAssetSimreadyResult:
-    """Result of making an asset simulation-ready."""
-
     output_path: Path
     transform_matrix: list[list[float]]
 
 
 @dataclass(frozen=True)
 class MakeTableSimreadyRequest:
-    """Request to prepare a generated table GLB for simulation placement."""
-
     input_path: Path
     output_path: Path
     input_up_axis: list[float] | None = None
@@ -52,7 +58,42 @@ class MakeTableSimreadyRequest:
 
 @dataclass(frozen=True)
 class MakeTableSimreadyResult:
-    """Result of making a table simulation-ready."""
-
     output_path: Path
     transform_matrix: list[list[float]]
+
+
+@dataclass(frozen=True)
+class MetricScaleObjectInput:
+    object_id: str
+    object_name: str
+    object_description: str
+    mesh_path: Path
+
+
+@dataclass(frozen=True)
+class EstimateMetricScalesRequest:
+    objects: list[MetricScaleObjectInput]
+    messages: list[dict[str, Any]]
+    schema: dict[str, Any]
+    llm: Any
+    context: str
+    method: str
+    step_name: str = "metric_scale"
+    raw_output_path: Path | None = None
+
+
+@dataclass(frozen=True)
+class EstimateMetricScalesResult:
+    status: str
+    object_scales: list[dict[str, Any]]
+    object_payload: list[dict[str, Any]]
+    raw_model_output: dict[str, Any] | None = None
+    reason: str = ""
+
+
+@dataclass(frozen=True)
+class GlobalMetricScaleRequest:
+    objects: list[dict[str, Any]]
+    object_scenes: list[tuple[str, Any]]
+    min_scale: float = 0.10
+    max_scale: float = 10.00
