--- embodichain/gen_sim/prompt2scene/agent_tools/managers/text_layout_manager/__init__.py
@@ -1,33 +1,0 @@
-# ----------------------------------------------------------------------------
-# Copyright (c) 2021-2026 DexForce Technology Co., Ltd.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ----------------------------------------------------------------------------
-
-from __future__ import annotations
-
-from embodichain.gen_sim.prompt2scene.agent_tools.managers.text_layout_manager.layout import (
-    _layout_text_objects_grid,
-)
-from embodichain.gen_sim.prompt2scene.agent_tools.managers.text_layout_manager.optimization import (
-    _optimize_text_layout_slp,
-)
-from embodichain.gen_sim.prompt2scene.agent_tools.managers.text_layout_manager.settle import (
-    settle_text_objects_to_ground,
-)
-
-__all__ = [
-    "_layout_text_objects_grid",
-    "_optimize_text_layout_slp",
-    "settle_text_objects_to_ground",
-]
