+++ embodichain/gen_sim/prompt2scene/agent_tools/tools/table_clutter_fit.py
@@ -17,24 +17,14 @@
 
 from __future__ import annotations
 
-import json
 import tempfile
 from pathlib import Path
 from typing import Any
 
 import numpy as np
 
-from embodichain.gen_sim.prompt2scene.utils.io import relative_path
-from embodichain.gen_sim.prompt2scene.agent_tools.managers.geometry_manager.scene_geometry import (
-    _copy_scene_with_transform,
-    _scene_to_mesh,
-    _z_up_to_glb_y_up_transform,
-    _detect_table_fit_support_quad,
-    _load_table_fit_scene_internal_z,
-    _table_fit_bounds_xy_manifest,
-    _table_fit_safe_positive_ratio,
-    _table_fit_scene_union_bounds,
-    _table_fit_uniform_xy_scale_transform,
+from embodichain.gen_sim.prompt2scene.agent_tools.managers.geometry_manager import (
+    GeometryManager,
 )
 from embodichain.gen_sim.prompt2scene.agent_tools.managers.simulation_manager import (
     SimulationManager,
@@ -65,7 +55,7 @@ def _gravity_settle_table_fit_internal_z_scene(
     with tempfile.TemporaryDirectory(prefix="p2s_table_fit_gravity_") as tmp:
         tmp_path = Path(tmp)
         pre_gravity = tmp_path / "table_pre_gravity.glb"
-        _copy_scene_with_transform(scene, z_to_y).export(pre_gravity)
+        GeometryManager.copy_scene_with_transform(scene, z_to_y).export(pre_gravity)
         result = sim.run_gravity_simulation(
             GravityDropRequest(
                 glb_path=pre_gravity,
@@ -78,20 +68,14 @@ def _gravity_settle_table_fit_internal_z_scene(
     return settled
 
 
-def _write_table_fit_json(path: Path, data: dict[str, Any]) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(
-        json.dumps(data, ensure_ascii=False, indent=2) + "\n",
-        encoding="utf-8",
-    )
-
-
 def fit_table_to_clutter(
     *,
     table_result: dict[str, Any],
     clutter_result: dict[str, Any],
     output_root: Path,
     output_dir: Path,
+    table_output_path: Path | None = None,
+    object_output_paths: dict[str, Path] | None = None,
     margin_cm: float = 10.0,
     support_occupancy_ratio: float = 0.80,
     object_coverage_percent: int | None = None,
@@ -115,6 +99,14 @@ def fit_table_to_clutter(
     output_root = output_root.expanduser().resolve()
     output_dir = output_dir.expanduser().resolve()
     output_dir.mkdir(parents=True, exist_ok=True)
+    if table_output_path is None:
+        table_output_path = output_dir / "table_fit_to_clutter.glb"
+    table_output_path = table_output_path.expanduser().resolve()
+    table_output_path.parent.mkdir(parents=True, exist_ok=True)
+    object_output_paths = {
+        str(key): path.expanduser().resolve()
+        for key, path in (object_output_paths or {}).items()
+    }
 
     # Resolve the table geometry.
     table_simready_path = _resolve_generated_path(
@@ -145,30 +137,37 @@ def fit_table_to_clutter(
     if not object_glb_paths:
         raise ValueError("No valid settled object GLBs for table fitting.")
 
-    z_to_y = _z_up_to_glb_y_up_transform()
+    z_to_y = GeometryManager.z_up_to_glb_y_up_transform()
     y_to_z = np.linalg.inv(z_to_y)
 
     # Load the table and detect its support surface.
-    table_scene = _load_table_fit_scene_internal_z(
+    table_scene = GeometryManager.load_table_fit_scene_internal_z(
         table_simready_path,
         trimesh=trimesh,
         y_to_z=y_to_z,
     )
-    table_mesh = _scene_to_mesh(table_scene, trimesh=trimesh)
+    table_mesh = GeometryManager.scene_to_mesh(table_scene, trimesh=trimesh)
     clutter_aabb = clutter_result.get("clutter_2d_aabb_cm") or {}
     clutter_size = clutter_aabb.get("size_xy", [1.0, 1.0])
     target_aspect = float(clutter_size[0]) / max(float(clutter_size[1]), 1.0e-6)
-    initial_support = _detect_table_fit_support_quad(
+    initial_support = GeometryManager.detect_table_fit_support_quad(
         table_mesh,
         target_aspect=target_aspect,
     )
 
     # Load the clutter scenes.
     clutter_scenes = [
-        (oid, _load_table_fit_scene_internal_z(path, trimesh=trimesh, y_to_z=y_to_z))
+        (
+            oid,
+            GeometryManager.load_table_fit_scene_internal_z(
+                path,
+                trimesh=trimesh,
+                y_to_z=y_to_z,
+            ),
+        )
         for oid, path in object_glb_paths
     ]
-    clutter_bounds = _table_fit_scene_union_bounds(
+    clutter_bounds = GeometryManager.table_fit_scene_union_bounds(
         [scene for _, scene in clutter_scenes],
         trimesh=trimesh,
     )
@@ -182,10 +181,16 @@ def fit_table_to_clutter(
     occupancy = float(np.clip(support_occupancy_ratio, 0.1, 1.0))
     required_size_cm = clutter_size_cm / occupancy + 2.0 * float(margin_cm)
     support_size_cm = np.asarray(initial_support["size_xy"], dtype=np.float64) * 100.0
-    scale_x = _table_fit_safe_positive_ratio(required_size_cm[0], support_size_cm[0])
-    scale_y = _table_fit_safe_positive_ratio(required_size_cm[1], support_size_cm[1])
+    scale_x = GeometryManager.table_fit_safe_positive_ratio(
+        required_size_cm[0],
+        support_size_cm[0],
+    )
+    scale_y = GeometryManager.table_fit_safe_positive_ratio(
+        required_size_cm[1],
+        support_size_cm[1],
+    )
     uniform_scale = max(scale_x, scale_y)
-    table_scale_transform = _table_fit_uniform_xy_scale_transform(
+    table_scale_transform = GeometryManager.table_fit_uniform_xy_scale_transform(
         center_xy=np.asarray(initial_support["center_xy"], dtype=np.float64),
         scale=uniform_scale,
     )
@@ -200,8 +205,8 @@ def fit_table_to_clutter(
         )
 
     # Reposition the table at the origin.
-    final_table_mesh = _scene_to_mesh(table_scene, trimesh=trimesh)
-    final_support = _detect_table_fit_support_quad(
+    final_table_mesh = GeometryManager.scene_to_mesh(table_scene, trimesh=trimesh)
+    final_support = GeometryManager.detect_table_fit_support_quad(
         final_table_mesh,
         target_aspect=float(required_size_cm[0] / max(required_size_cm[1], 1.0e-6)),
     )
@@ -218,7 +223,10 @@ def fit_table_to_clutter(
     # Use the highest point of the table mesh (after scaling + gravity + shift)
     # rather than the support-plane mean Z, so that thin objects sit above the
     # actual geometry even when the tabletop has slight unevenness.
-    _table_mesh_after_shift = _scene_to_mesh(table_scene, trimesh=trimesh)
+    _table_mesh_after_shift = GeometryManager.scene_to_mesh(
+        table_scene,
+        trimesh=trimesh,
+    )
     _table_max_z = float(
         np.asarray(_table_mesh_after_shift.bounds, dtype=np.float64)[1, 2]
     )
@@ -227,13 +235,13 @@ def fit_table_to_clutter(
     # Place the objects on the table.
     placed_objects: list[dict[str, Any]] = []
     shifted_clutter: list[tuple[str, Any]] = []
-    clutter_after = _table_fit_scene_union_bounds(
+    clutter_after = GeometryManager.table_fit_scene_union_bounds(
         [scene for _, scene in clutter_scenes],
         trimesh=trimesh,
     )
     clutter_center_xy = 0.5 * (clutter_after[0, :2] + clutter_after[1, :2])
     for oid, scene in clutter_scenes:
-        obj_mesh = _scene_to_mesh(scene, trimesh=trimesh)
+        obj_mesh = GeometryManager.scene_to_mesh(scene, trimesh=trimesh)
         obj_bounds = np.asarray(obj_mesh.bounds, dtype=np.float64)
         obj_bottom_z = float(obj_bounds[0, 2])
         obj_shift = np.eye(4, dtype=np.float64)
@@ -246,16 +254,18 @@ def fit_table_to_clutter(
         shifted_clutter.append((oid, scene))
 
     # Export the fitted table and placed objects.
-    final_table_path = output_dir / "table_fit_to_clutter.glb"
-    _copy_scene_with_transform(table_scene, z_to_y).export(final_table_path)
+    GeometryManager.copy_scene_with_transform(table_scene, z_to_y).export(
+        table_output_path
+    )
 
     for oid, scene in shifted_clutter:
-        object_path = output_dir / f"{oid}_on_table.glb"
-        _copy_scene_with_transform(scene, z_to_y).export(object_path)
+        object_path = object_output_paths.get(oid, output_dir / f"{oid}_on_table.glb")
+        object_path.parent.mkdir(parents=True, exist_ok=True)
+        GeometryManager.copy_scene_with_transform(scene, z_to_y).export(object_path)
         # Compute world-space AABB bottom-centre (sim Z-up coords) before
         # the scene is converted to GLB Y-up for export.  This is the
         # reference position that gym_export uses to derive ``init_pos``.
-        _placed_mesh = _scene_to_mesh(scene, trimesh=trimesh)
+        _placed_mesh = GeometryManager.scene_to_mesh(scene, trimesh=trimesh)
         _placed_b = np.asarray(_placed_mesh.bounds, dtype=np.float64)
         world_aabb_bottom_center = [
             float(0.5 * (_placed_b[0, 0] + _placed_b[1, 0])),
@@ -270,12 +280,11 @@ def fit_table_to_clutter(
             }
         )
 
-    # Write the fit manifest.
-    final_clutter_bounds = _table_fit_scene_union_bounds(
+    final_clutter_bounds = GeometryManager.table_fit_scene_union_bounds(
         [scene for _, scene in shifted_clutter],
         trimesh=trimesh,
     )
-    final_clutter_aabb_cm = _table_fit_bounds_xy_manifest(
+    final_clutter_aabb_cm = GeometryManager.table_fit_bounds_xy_manifest(
         final_clutter_bounds,
         unit_scale=100.0,
     )
@@ -295,7 +304,7 @@ def fit_table_to_clutter(
         "status": "ok",
         "output_dir": str(output_dir),
         "table_simready_path": str(table_simready_path),
-        "table_output_path": str(final_table_path),
+        "table_output_path": str(table_output_path),
         "objects": placed_objects,
         "margin_cm": margin_cm,
         "support_occupancy_ratio": occupancy,
@@ -319,9 +328,4 @@ def fit_table_to_clutter(
             <= float(np.asarray(final_support_centered["size_xy"])[1] * 100.0),
         },
     }
-    manifest_path = output_dir / "table_fit_to_clutter_manifest.json"
-    _write_table_fit_json(manifest_path, manifest)
-    return {
-        "status": "ok",
-        "manifest_path": relative_path(manifest_path, output_root),
-    }
+    return manifest
