+++ embodichain/gen_sim/prompt2scene/agent_tools/tools/table_fit_scene.py
@@ -20,12 +20,27 @@ import traceback
 from pathlib import Path
 from typing import Any
 
-from embodichain.gen_sim.prompt2scene.agent_tools.managers.table_clutter_fit_manager import (
+from embodichain.gen_sim.prompt2scene.agent_tools.tools.table_clutter_fit import (
     fit_table_to_clutter,
 )
+from embodichain.gen_sim.prompt2scene.utils.io import relative_path, write_json
 from embodichain.gen_sim.prompt2scene.utils.log import log_info, log_warning
 
 __all__ = ["fit_image_scene_table", "fit_text_scene_table"]
+
+
+def _table_fit_output_paths(
+    *,
+    output_dir: Path,
+    object_ids: list[str],
+) -> tuple[Path, dict[str, Path]]:
+    return (
+        output_dir / "table_fit_to_clutter.glb",
+        {
+            object_id: output_dir / f"{object_id}_on_table.glb"
+            for object_id in object_ids
+        },
+    )
 
 
 def fit_text_scene_table(
@@ -37,13 +52,30 @@ def fit_text_scene_table(
 ) -> dict[str, Any]:
     """Fit the text-scene table and convert failures to result data."""
     try:
-        result = fit_table_to_clutter(
+        object_ids = [
+            str(item["id"])
+            for item in clutter_layout_result.get("objects", [])
+            if isinstance(item, dict) and item.get("id") and item.get("status") == "ok"
+        ]
+        table_output_path, object_output_paths = _table_fit_output_paths(
+            output_dir=output_dir,
+            object_ids=object_ids,
+        )
+        manifest = fit_table_to_clutter(
             table_result=table_result,
             clutter_result=clutter_layout_result,
             output_root=output_root,
             output_dir=output_dir,
+            table_output_path=table_output_path,
+            object_output_paths=object_output_paths,
             object_coverage_percent=table_result.get("object_coverage_percent"),
         )
+        manifest_path = output_dir / "table_fit_to_clutter_manifest.json"
+        write_json(manifest_path, manifest)
+        result = {
+            "status": "ok",
+            "manifest_path": relative_path(str(manifest_path), output_root),
+        }
         log_info(f"text table fit completed status={result.get('status')}")
         return result
     except Exception as exc:
@@ -76,6 +108,17 @@ def fit_image_scene_table(
         }
 
     try:
+        object_ids = [
+            str(item["id"])
+            for item in generated_objects
+            if isinstance(item, dict)
+            and item.get("id")
+            and item.get("aligned_geometry_path")
+        ]
+        table_output_path, object_output_paths = _table_fit_output_paths(
+            output_dir=output_dir,
+            object_ids=object_ids,
+        )
         clutter_result = {
             "clutter_2d_aabb_cm": alignment_result.get(
                 "final_clutter_2d_aabb_cm"
@@ -90,13 +133,21 @@ def fit_image_scene_table(
                 if item.get("id") and item.get("aligned_geometry_path")
             ],
         }
-        result = fit_table_to_clutter(
+        manifest = fit_table_to_clutter(
             table_result=generated_table,
             clutter_result=clutter_result,
             output_root=output_root,
             output_dir=output_dir,
+            table_output_path=table_output_path,
+            object_output_paths=object_output_paths,
             object_coverage_percent=generated_table.get("object_coverage_percent"),
         )
+        manifest_path = output_dir / "table_fit_to_clutter_manifest.json"
+        write_json(manifest_path, manifest)
+        result = {
+            "status": "ok",
+            "manifest_path": relative_path(str(manifest_path), output_root),
+        }
         log_info(f"image table fit completed status={result.get('status')}")
         return result
     except Exception as exc:
