+++ embodichain/gen_sim/prompt2scene/workflows/image_relations/nodes.py
@@ -19,9 +19,9 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Any
 
-from embodichain.gen_sim.prompt2scene.agent_tools.clients.image_segmentation_client import (
-    decode_rle_mask,
-    draw_numbered_masks,
+from embodichain.gen_sim.prompt2scene.agent_tools.tools.image_segment_filter import (
+    filter_group_segments_with_vlm,
+    filter_segments_with_vlm,
 )
 from embodichain.gen_sim.prompt2scene.workflows.image_relations.schema import (
     ImageAssetSegment,
@@ -29,8 +29,7 @@ from embodichain.gen_sim.prompt2scene.workflows.image_relations.schema import (
     ImageRelationSpec,
 )
 from embodichain.gen_sim.prompt2scene.workflows.request import InputKind
-from embodichain.gen_sim.prompt2scene.workflows.image_relations.schema import (
-    FILTER_EXTRA_INSTANCES_JSON_SCHEMA,
+from embodichain.gen_sim.prompt2scene.prompts.schemas import (
     SPATIAL_LAYOUT_JSON_SCHEMA,
 )
 from embodichain.gen_sim.prompt2scene.utils import (
@@ -48,23 +47,24 @@ from embodichain.gen_sim.prompt2scene.workflows.image_relations.utils import (
     asset_bbox_label,
     draw_labeled_bboxes,
     expand_asset_ids,
-    filter_group_segments_with_vlm,
-    filter_segments_with_vlm,
     merge_non_overlapping_segments,
     prompt_text,
     path_token,
     require_image_path,
+    segment_area,
     segment_prompt,
     segments_from_response,
+    select_largest_table_segment,
+    table_segmentation_prompts,
+    write_table_candidate_debug_image,
 )
-from embodichain.gen_sim.prompt2scene.workflows.image_relations.prompts import (
-    build_filter_extra_instances_messages,
+from embodichain.gen_sim.prompt2scene.prompts.builders import (
     build_spatial_layout_messages,
 )
 from embodichain.gen_sim.prompt2scene.workflows.image_relations.state import (
     ImageRelationsState,
 )
-from embodichain.gen_sim.prompt2scene.workflows.llm_output import (
+from embodichain.gen_sim.prompt2scene.llms.llm_output import (
     call_structured_json_model_step,
     is_model_output_error,
 )
@@ -140,7 +140,20 @@ def call_vlm_filter_initial_segments_node(
     llm: Any,
 ) -> dict[str, object]:
     """Ask VLM to remove wrong masks from initial name-based SAM3 output."""
-    return filter_segments_with_vlm(state=state, llm=llm, stage="initial")
+    image_path = require_image_path(state)
+    art = WorkflowArtifactWriter(state["output_root"], IMAGE_SEGMENTS_STEP)
+    return filter_segments_with_vlm(
+        llm=llm,
+        image_path=image_path,
+        step_name=IMAGE_SEGMENTS_STEP,
+        segment_groups=state["segment_groups"],
+        attempt_count=state["attempt_count"],
+        errors=state["errors"],
+        stage="initial",
+        next_debug_round_name=art.next_debug_round_name,
+        debug_round_dir=art.debug_round_dir,
+        write_debug_json=art.write_debug_round_json,
+    )
 def retry_missing_by_candidates_node(
     state: ImageRelationsState,
     *,
@@ -167,13 +180,21 @@ def retry_missing_by_candidates_node(
                 response=response,
                 source_prompt=prompt,
             )
+            stage_label = f"fallback_{path_token(prompt)}"
+            round_name_inner = artifact_writer.next_debug_round_name(
+                label=f"{stage_label}_{group['name']}"
+            )
+            round_dir_inner = artifact_writer.debug_round_dir(round_name_inner)
             new_segments = filter_group_segments_with_vlm(
                 llm=llm,
                 image_path=image_path,
-                artifact_writer=artifact_writer,
+                step_name=IMAGE_SEGMENTS_STEP,
                 group=group,
                 segments=new_segments,
-                stage=f"fallback_{path_token(prompt)}",
+                stage=stage_label,
+                debug_round_name=round_name_inner,
+                debug_round_dir=round_dir_inner,
+                write_debug_json=artifact_writer.write_debug_round_json,
             )
             segments = merge_non_overlapping_segments(
                 existing=segments,
@@ -324,7 +345,7 @@ def segment_table_node(
     }
     segments: list[dict[str, Any]] = []
 
-    for prompt in _table_segmentation_prompts(group):
+    for prompt in table_segmentation_prompts(group):
         if len(segments) >= 1:
             break
         response = segment_prompt(image_path=image_path, prompt=prompt)
@@ -334,14 +355,14 @@ def segment_table_node(
             response=response,
             source_prompt=prompt,
         )
-        _write_table_candidate_debug_image(
+        write_table_candidate_debug_image(
             image_path=image_path,
             artifact_writer=artifact_writer,
             group=group,
             segments=new_segments,
             stage=f"table_{path_token(prompt)}",
         )
-        selected_segment = _select_largest_table_segment(new_segments)
+        selected_segment = select_largest_table_segment(new_segments)
         if selected_segment is not None:
             segments = [selected_segment]
 
@@ -387,66 +408,9 @@ def segment_table_node(
     return {"image_relations": updated_image_relations}
 
 
-def _table_segmentation_prompts(group: dict[str, Any]) -> list[str]:
-    """Return table/support segmentation prompts in object-style fallback order."""
-    prompts = [prompt_text(group["name"])]
-    for candidate_name in group["class_candidate"][1:]:
-        prompts.append(prompt_text(candidate_name))
-    description_prompt = str(group.get("description") or "").strip()
-    if description_prompt:
-        prompts.append(description_prompt)
 
-    unique_prompts: list[str] = []
-    for prompt in prompts:
-        if prompt and prompt not in unique_prompts:
-            unique_prompts.append(prompt)
-    return unique_prompts
-
-
-def _write_table_candidate_debug_image(
-    *,
-    image_path: Path,
-    artifact_writer: WorkflowArtifactWriter,
-    group: dict[str, Any],
-    segments: list[dict[str, Any]],
-    stage: str,
-) -> None:
-    """Write table/support candidate mask debug image without VLM filtering."""
-    if not segments:
-        return
-    round_name = artifact_writer.next_debug_round_name(label=f"{stage}_{group['name']}")
-    round_dir = artifact_writer.debug_round_dir(round_name)
-    debug_image_path = draw_numbered_masks(
-        image_path=image_path,
-        segments=segments,
-        output_path=round_dir / "mask.png",
-    )
-    group["debug_images"] = append_unique(
-        group["debug_images"],
-        str(debug_image_path),
-    )
-
-
-def _select_largest_table_segment(
-    segments: list[dict[str, Any]],
-) -> dict[str, Any] | None:
-    """Select the largest SAM3 table/support candidate without VLM filtering."""
-    if not segments:
-        return None
-    return max(segments, key=_segment_area)
-
-
-def _segment_area(segment: dict[str, Any]) -> float:
-    mask_rle = segment.get("mask_rle")
-    if mask_rle is not None:
-        try:
-            mask = decode_rle_mask(mask_rle).convert("L")
-            histogram = mask.histogram()
-            return float(sum(count for value, count in enumerate(histogram) if value))
-        except Exception:
-            pass
-    x1, y1, x2, y2 = segment["bbox_xyxy"]
-    return max(0.0, float(x2) - float(x1)) * max(0.0, float(y2) - float(y1))
+    artifact_writer.write_step_result(updated_image_relations.to_segmentation_manifest())
+    return {"image_relations": updated_image_relations}
 
 
 def call_vlm_spatial_layout_node(
@@ -483,11 +447,11 @@ def call_vlm_spatial_layout_node(
             schema=SPATIAL_LAYOUT_JSON_SCHEMA,
             messages=messages,
             context="Image spatial layout",
-            step_name=IMAGE_SPATIAL_RELATIONS_STEP,
-            output_root=None,
+            
+            
             attempt_count=attempt_count,
-            raw_output_label="spatial_layout",
-            artifact_writer=artifact_writer,
+            
+            
         )
         updated_image_relations = apply_spatial_layout_output(
             image_relations=image_relations,
@@ -496,7 +460,9 @@ def call_vlm_spatial_layout_node(
         artifact_writer.write_step_result(updated_image_relations.to_spatial_manifest())
     except Exception as exc:
         if is_model_output_error(exc) or isinstance(exc, ValueError):
-            error = format_attempt_error("Image relations spatial layout", attempt_count, exc)
+            error = format_attempt_error(
+                "Image relations spatial layout", attempt_count, exc
+            )
             log.log_warning(error)
             return {
                 "attempt_count": attempt_count,
