+++ embodichain/gen_sim/prompt2scene/workflows/image_relations/utils.py
@@ -25,21 +25,23 @@ from embodichain.gen_sim.prompt2scene.agent_tools.clients.image_segmentation_cli
     ImageSegmentationServerRequest,
     ImageSegmentationServerResponse,
     bbox_iou,
+    decode_rle_mask,
     draw_labeled_bboxes,
     draw_numbered_masks,
     is_usable_segmentation_candidate,
     sort_segments_by_bbox,
 )
+from embodichain.gen_sim.prompt2scene.prompts.schemas import (
+    SPATIAL_LAYOUT_JSON_SCHEMA,
+)
 from embodichain.gen_sim.prompt2scene.workflows.image_relations.schema import (
-    FILTER_EXTRA_INSTANCES_JSON_SCHEMA,
     ImageAnchor,
     ImageAssetLayout,
     ImageAssetSegment,
     ImageRelationGroup,
     ImageRelationSpec,
-    SPATIAL_LAYOUT_JSON_SCHEMA,
 )
-from embodichain.gen_sim.prompt2scene.workflows.spatial import (
+from embodichain.gen_sim.prompt2scene.agent_tools.tools.spatial_relations import (
     GRID_VALUES,
     validate_exact_asset_id_coverage,
 )
@@ -47,19 +49,13 @@ from embodichain.gen_sim.prompt2scene.utils import log_api_request_start, log
 from embodichain.gen_sim.prompt2scene.workflows.artifact_writer import (
     IMAGE_SEGMENTS_STEP,
     IMAGE_SPATIAL_RELATIONS_STEP,
-    RAW_MODEL_OUTPUT_FILENAME,
     WorkflowArtifactWriter,
 )
-from embodichain.gen_sim.prompt2scene.workflows.image_relations.prompts import (
-    build_filter_extra_instances_messages,
+from embodichain.gen_sim.prompt2scene.prompts.builders import (
     build_spatial_layout_messages,
 )
-from embodichain.gen_sim.prompt2scene.workflows.llm_output import (
+from embodichain.gen_sim.prompt2scene.llms.llm_output import (
     call_structured_json_model_step,
-    is_model_output_error,
-)
-from embodichain.gen_sim.prompt2scene.workflows.stage_errors import (
-    format_attempt_error,
 )
 
 __all__ = [
@@ -68,21 +64,22 @@ __all__ = [
     "append_unique",
     "apply_spatial_layout_output",
     "asset_bbox_label",
-    "expand_asset_ids",
-    "filter_group_segments_with_vlm",
-    "filter_segments_with_vlm",
-    "merge_non_overlapping_segments",
     "draw_labeled_bboxes",
+    "expand_asset_ids",
+    "merge_non_overlapping_segments",
     "parse_anchor",
     "parse_asset_states",
     "parse_order_groups",
     "path_token",
     "prompt_text",
-    "remove_extra_numbered_segments",
     "require_image_path",
+    "segment_area",
     "segment_prompt",
     "segments_from_response",
+    "select_largest_table_segment",
     "sort_segments_by_bbox",
+    "table_segmentation_prompts",
+    "write_table_candidate_debug_image",
 ]
 
 MAX_SEGMENT_RETRIES = 1
@@ -297,121 +294,68 @@ def parse_asset_states(
     return state_by_asset_id
 
 
-def filter_group_segments_with_vlm(
+def table_segmentation_prompts(group: dict[str, Any]) -> list[str]:
+    """Return table/support segmentation prompts in object-style fallback order."""
+    prompts = [prompt_text(group["name"])]
+    for candidate_name in group["class_candidate"][1:]:
+        prompts.append(prompt_text(candidate_name))
+    description_prompt = str(group.get("description") or "").strip()
+    if description_prompt:
+        prompts.append(description_prompt)
+
+    unique_prompts: list[str] = []
+    for prompt in prompts:
+        if prompt and prompt not in unique_prompts:
+            unique_prompts.append(prompt)
+    return unique_prompts
+
+
+def write_table_candidate_debug_image(
     *,
-    llm: Any,
     image_path: Path,
     artifact_writer: WorkflowArtifactWriter,
     group: dict[str, Any],
     segments: list[dict[str, Any]],
     stage: str,
-) -> list[dict[str, Any]]:
-    """Ask VLM to remove wrong or duplicate instances from one SAM3 result."""
-    segments = sort_segments_by_bbox(segments)
+) -> None:
+    """Write table/support candidate mask debug image without VLM filtering."""
     if not segments:
-        return segments
-
-    round_name = artifact_writer.next_debug_round_name(label=f"{stage}_{group['name']}")
+        return
+    round_name = artifact_writer.next_debug_round_name(
+        label=f"{stage}_{group['name']}"
+    )
     round_dir = artifact_writer.debug_round_dir(round_name)
     debug_image_path = draw_numbered_masks(
         image_path=image_path,
         segments=segments,
         output_path=round_dir / "mask.png",
     )
-    group["debug_images"] = append_unique(
-        group["debug_images"],
-        str(debug_image_path),
-    )
-    log_api_request_start(
-        step=IMAGE_SEGMENTS_STEP,
-        request=f"vlm_filter_{stage}",
-        debug_image=str(debug_image_path),
-    )
-    messages = build_filter_extra_instances_messages(
-        debug_image_path=debug_image_path,
-        name=group["name"],
-        description=group["description"],
-        expected_count=group["expected_count"],
-        class_candidate=group["class_candidate"],
-    )
-    raw_model_output = call_structured_json_model_step(
-        llm=llm,
-        schema=FILTER_EXTRA_INSTANCES_JSON_SCHEMA,
-        messages=messages,
-        context=f"Image relation {stage} segmentation filtering",
-        step_name=IMAGE_SEGMENTS_STEP,
-        output_root=None,
-        attempt_count=0,
-        raw_output_writer=lambda payload: artifact_writer.write_debug_round_json(
-            round_name=round_name,
-            filename=RAW_MODEL_OUTPUT_FILENAME,
-            payload=payload,
-        ),
-    )
-    return remove_extra_numbered_segments(
-        segments=segments,
-        raw_model_output=raw_model_output,
-    )
+    debug_images = list(group.get("debug_images") or [])
+    if str(debug_image_path) not in debug_images:
+        debug_images.append(str(debug_image_path))
+    group["debug_images"] = debug_images
 
 
-def filter_segments_with_vlm(
-    *,
-    state: dict[str, Any],
-    llm: Any,
-    stage: str,
-) -> dict[str, object]:
-    """Filter all segment groups with VLM and return an updated state patch."""
-    segment_groups = []
-    attempt_count = state["attempt_count"] + 1
-    image_path = require_image_path(state)
-    artifact_writer = WorkflowArtifactWriter(state["output_root"], IMAGE_SEGMENTS_STEP)
-
-    try:
-        for group in state["segment_groups"]:
-            group = dict(group)
-            group["segments"] = filter_group_segments_with_vlm(
-                llm=llm,
-                image_path=image_path,
-                artifact_writer=artifact_writer,
-                group=group,
-                segments=group["segments"],
-                stage=stage,
-            )
-            segment_groups.append(group)
-    except Exception as exc:
-        if is_model_output_error(exc) or isinstance(exc, ValueError):
-            error = format_attempt_error("Image relations VLM filter", attempt_count, exc)
-            log.log_warning(error)
-            return {
-                "attempt_count": attempt_count,
-                "last_error": error,
-                "errors": state["errors"] + [error],
-            }
-        raise
-
-    return {
-        "attempt_count": attempt_count,
-        "segment_groups": segment_groups,
-        "last_error": None,
-    }
-
-
-def remove_extra_numbered_segments(
-    *,
+def select_largest_table_segment(
     segments: list[dict[str, Any]],
-    raw_model_output: dict[str, Any],
-) -> list[dict[str, Any]]:
-    """Remove numbered masks flagged as extra by the VLM."""
-    extra_numbers = raw_model_output.get("extra_instance_numbers")
-    if not isinstance(extra_numbers, list):
-        raise ValueError("extra_instance_numbers must be a list.")
-    extra_indices = {int(number) - 1 for number in extra_numbers}
-    if any(index < 0 or index >= len(segments) for index in extra_indices):
-        raise ValueError("VLM returned an out-of-range extra mask number.")
-    kept = [
-        segment for index, segment in enumerate(segments) if index not in extra_indices
-    ]
-    return kept
+) -> dict[str, Any] | None:
+    """Select the largest SAM3 table/support candidate without VLM filtering."""
+    if not segments:
+        return None
+    return max(segments, key=segment_area)
+
+
+def segment_area(segment: dict[str, Any]) -> float:
+    mask_rle = segment.get("mask_rle")
+    if mask_rle is not None:
+        try:
+            mask = decode_rle_mask(mask_rle).convert("L")
+            histogram = mask.histogram()
+            return float(sum(count for value, count in enumerate(histogram) if value))
+        except Exception:
+            pass
+    x1, y1, x2, y2 = segment["bbox_xyxy"]
+    return max(0.0, float(x2) - float(x1)) * max(0.0, float(y2) - float(y1))
 
 
 def merge_non_overlapping_segments(
@@ -428,7 +372,8 @@ def merge_non_overlapping_segments(
         if len(merged) >= limit:
             break
         if all(
-            bbox_iou(segment["bbox_xyxy"], other["bbox_xyxy"]) < OVERLAP_IOU_THRESHOLD
+            bbox_iou(segment["bbox_xyxy"], other["bbox_xyxy"])
+            < OVERLAP_IOU_THRESHOLD
             for other in merged
         ):
             merged.append(segment)
